+++ pkg/R/zzz.R
@@ -74,6 +74,3 @@ NULL
 
 ## create the logging environment
 logging.options <- new.env()
-
-## initialize the module
-logReset()